+++ shaders/program/voxy_translucent.glsl
@@ -1,5 +1,5 @@
 #undef SHADOWS_TYPE
-#define SHADOWS_TYPE 0
+#define SHADOWS_TYPE 1
 
 layout(location = 0) out vec4 albedoOut;
 layout(location = 1) out vec4 auxDataOut;
@@ -58,6 +58,11 @@ void voxy_emitFragment(VoxyFragmentParameters parameters) {
 	
 	// vsh lighting
 	doVshLighting(lmcoord, glcolor, viewPos, normal, worldNormal);
+	#if SHADOWS_TYPE == 1
+		lmcoord.y *= 0.875 + 0.125 * step(0.96, lmcoord.y); // add fake shadows
+	#else
+		lmcoord.y *= 0.75 + 0.25 * step(0.96, lmcoord.y); // add fake shadows
+	#endif
 	
 	
 	// main color