+++ changelog.md
@@ -3,6 +3,7 @@
   - Increased 'Reflection Iterations' from 45 to 60
   - Changed default 'Shadow Map Distance Cutoff' value from x2.0 to x1.0
   - Performance compared to last version can be up to 7-9% better (with 16 chunks render distance + voxy + 1080p)
+  - Added fake shadows to LOD terrain
   - Added setting 'Particles Brightness'
   - Added settings 'Leaves Side Shading' and 'Leaves Ao Amount'
   - Added settings 'End Sky Brightness' and 'End Static Brightness'